Convert between NE and DN glycan graphs

This function converts a Node-Edge (NE) glycan graph into a Dual-Node (DN) glycan graph, or vice versa. In an NE glycan graph, nodes represent monosaccharides and edges represent linkages. In a DN glycan graph, both monosaccharides and linkages are nodes, and they alternate. See as_glycan_graph() for details.

Usage

convert_graph_mode(glycan, to, strict = TRUE)

Arguments

glycan

A glycan graph.

to

The target graph mode, either "ne" or "dn".

strict

If TRUE, an error will be thrown if the glycan is already in the target mode. If FALSE, the glycan will be returned as is. Default is TRUE.

Value

The converted glycan graph object.

Examples

(glycan <- n_glycan_core(mode = "ne"))
#> Glycan Graph (NE)
#> GlcNAc: 2, Man: 3
#> ------------------
#> GlcNAc (?1-)
#> └─GlcNAc (b1-4)
#>   └─Man (b1-4)
#>     ├─Man (a1-3)
#>     └─Man (a1-6)

# Convert to DN mode
convert_graph_mode(glycan, to = "dn")
#> Glycan Graph (DN)
#> GlcNAc: 2, Man: 3
#> ------------------
#> GlcNAc (?1-)
#> └─b1-4
#>   └─GlcNAc
#>     └─b1-4
#>       └─Man
#>         ├─a1-3
#>         │ └─Man
#>         └─a1-6
#>           └─Man

# Converting to NE mode will raise an error
try(convert_graph_mode(glycan, to = "ne"))
#> Error in convert_graph_mode(glycan, to = "ne") : 
#>   The glycan is already in "NE" mode.

# Set `strict` to `FALSE` to return the glycan unchanged
convert_graph_mode(glycan, to = "ne", strict = FALSE)
#> Glycan Graph (NE)
#> GlcNAc: 2, Man: 3
#> ------------------
#> GlcNAc (?1-)
#> └─GlcNAc (b1-4)
#>   └─Man (b1-4)
#>     ├─Man (a1-3)
#>     └─Man (a1-6)