Example Glycan Graphs

Create example glycan graphs for testing and demonstration. Includes N-glycan core and O-glycan core 1 and core 2.

Usage

n_glycan_core(mode = "ne", linkage = TRUE, mono_type = "concrete")

o_glycan_core_1(mode = "ne", linkage = TRUE, mono_type = "concrete")

o_glycan_core_2(mode = "ne", linkage = TRUE, mono_type = "concrete")

Arguments

mode

A character string, either "ne" or "dn". See as_glycan_graph() for details.

linkage

A logical indicating whether to include linkages (e.g. "b1-4"). Default is TRUE.

mono_type

A character string specifying the type of monosaccharides. Can be "simple" (H, N, F, S), "generic" (Hex, HexNAc, dHex, NeuAc, etc.), or "concrete" (Man, Gal, HexNAc, Fuc, etc.). Default is "concrete".

Value

A glycan graph object.

N-Glycan Core

N-Glycans are branched oligosaccharides that are bound, most commonly, via GlcNAc to an Asn residue of the protein backbone. A common motif of all N-glycans is the chitobiose core, composed of three mannose and two GlcNAc moieties, which is commonly attached to the protein backbone via GlcNAc. The mannose residue is branched and connected via α1,3- and α1,6-glycosidic linkages to the two other mannose building blocks.

    Man
  a1-6 \   b1-4      b1-4
        Man -- GlcNAc -- GlcNAc
  a1-3 /
    Man

O-Glycan Core

O-Glycans are highly abundant in extracellular proteins. Generally, O-glycans are extended following four major core structures: core 1, core 2, core 3, and core 4. The first two are by far the most common core structures in O-glycosylation and are found throughout the body.

core 1:

    GalNAc
   / b1-3
Gal

core 2:

GlcNAc
      \ b1-6
       GalNAc
      / b1-3
   Gal

Examples

print(n_glycan_core(mode = "ne"), verbose = TRUE)
#> Glycan Graph (NE)
#> GlcNAc: 2, Man: 3
#> ------------------
#> GlcNAc (?1-)
#> └─GlcNAc (b1-4)
#>   └─Man (b1-4)
#>     ├─Man (a1-3)
#>     └─Man (a1-6)
print(o_glycan_core_1(mode = "dn"), verbose = TRUE)
#> Glycan Graph (DN)
#> Gal: 1, GalNAc: 1
#> ------------------
#> GalNAc (a1-)
#> └─b1-3
#>   └─Gal