Convert igraph Graph to Glycan Structure

A glycan structure is a subclass of igraph graph with additional constraints. This function checks these constraints and append the S3 class glycan_structure to the graph object.

Generally, it is not recommended to create a glycan structure using this function manually. Use the glyparse package to generate a glycan structure from a structure string.

A glycan structure is a directed modeling of a glycan structure, where nodes are monosaccharides and edges are linkages.

Usage

as_glycan_structure(graph)

is_glycan_structure(x)

Arguments

graph

An igraph graph object.

x

An object to check.

Value

A glycan structure object.

S3 Classes

A glycan structure is merely an igraph graph with an additional S3 class glycan_structure. Therefore, sloop::s3_class() of a glycan structure is c("glycan_structure", "igraph").

Constraints:

The graph must be directed and an outward tree (reducing end as root).

The graph must have a graph attribute anomer, in the form of "a1".

The graph must have a graph attribute alditol, a logical value.

The graph must have a vertex attribute mono for monosaccharide names.

The graph must have a vertex attribute sub for substituents.

The graph must have an edge attribute linkage for linkages.

Monosaccharide name must be known, either generic (Hex, HexNAc, etc.) or concrete (Glc, Gal, etc.), but not a mixture of both. NA is not allowed.

Substituent must be valid, in the form of "xY", where x is position and Y is substituent name, e.g. "2Ac", "3S", etc.

Linkages must be valid, in the form of "a/bX-Y", where X and Y are integers, e.g. "b1-4", "a2-3", etc. NA is not allowed.

Examples

library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union

# A simple glycan structure: GlcNAc(b1-4)GlcNAc
graph <- make_graph(~ 1-+2)  # 1 and 2 are monosaccharides
V(graph)$mono <- c("GlcNAc", "GlcNAc")
V(graph)$sub <- ""
E(graph)$linkage <- "b1-4"
graph$anomer <- "a1"
graph$alditol <- FALSE
as_glycan_structure(graph)
#> Glycan Structure
#> GlcNAc(2)
#> ------------------
#> GlcNAc (a1-)
#> └─GlcNAc (b1-4)